backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-cors-headers==4.4.0
Pillow==10.4.0
openpyxl==3.1.5
mysqlclient==2.2.4
social-auth-app-django==5.4.0
social-auth-core[wechat]==4.8.5
